refactor: hide the simulator files from import
