+++ src/orquestra/integrations/cirq/simulator/__init__.py
@@ -1,9 +1,9 @@
 ################################################################################
 # © Copyright 2021-2022 Zapata Computing Inc.
 ################################################################################
-from .cirq_simulator import CirqSimulator
+from ._cirq_simulator import CirqSimulator
 
 try:
-    from .qsim_simulator import QSimSimulator
+    from ._qsim_simulator import QSimSimulator
 except ModuleNotFoundError:
     pass
