+++ src/orquestra/integrations/custatevec/simulator/__init__.py
@@ -1,1 +1,1 @@
-from .custatevec import CuStateVecSimulator
+from ._custatevec import CuStateVecSimulator
